requirements.txt
Flask==2.0.1
numpy==1.17.4
Pillow==8.2.0
